+++ EtoE/wolvez2023_pkg/Wolvez2023/AR_powerplanner.py
@@ -1,6 +1,8 @@
 import numpy as np
 import cv2
 
+from numpy import sin, cos, arccos
+
 class ARPowerPlanner():
 
     #速度の設定
@@ -10,7 +12,7 @@ class ARPowerPlanner():
     def __init__(self):
         self.arm_id = "1"
         # 各マーカーに対するxg,yg,zg
-        self.marker_goal = {"3":[0.01,0.01,0.01],"4":[0.008,0.0,-0.006],"5":[-0.01,0,0],"6":[-0.008,0.01,-0.006],"10":[1,1,1]}
+        self.marker_goal = {"2":[0.0,0.0,0.01],"3":[0,0.001,0.038],"4":[0,0.042,-0.005],"5":[0,0.042,-0.005],"6":[0,0.042,-0.005],"7":[0,0.042,-0.005],"10":[1,1,1],"68":[0,0,-0.06]}
         # 参照するマーカーの優先度順
         self.marker_ref = ["10","5","3","6","4"]
 
@@ -28,32 +30,34 @@ class ARPowerPlanner():
         roll = ar_info[id]['roll']
         pitch = ar_info[id]['pitch']
         yaw = ar_info[id]['yaw']
+        rvec = ar_info[id]['rvec']
         bias = self.marker_goal[id] # マーカーからゴールまでのベクトル
 
-        goal = self.rot_vec(roll,pitch,yaw,bias)+np.array([[x],[y],[z]])
+        goal = self.rot_vec(rvec,bias)+np.array([[x],[y],[z]])
         goal= goal.T
         
         return goal.reshape(1,3)
     
-    def rot_vec(self,roll,pitch,yaw,vec):
-        rvec = np.array([roll, pitch, yaw]) 
-        rvec_matrix = cv2.Rodrigues(rvec) # rodorigues
+    def rot_vec(self,rvec,vec):
+        #rvec_norm = np.linalg.norm(rvec)
+        rvec_matrix = cv2.Rodrigues(rvec)
         rvec_matrix = rvec_matrix[0] # rodoriguesから抜き出し
-
-        g = np.dot(rvec_matrix,vec).reshape(-1, 1)
+        g = np.dot((rvec_matrix),vec).reshape(-1, 1)
         return g
 
     def goalvec_maker(self,ar_info,goal_point,connecting_state,id):
+        self.connecting_state = connecting_state
         if connecting_state == 0:
             if self.arm_id in ar_info.keys():
                 marker_1 = np.array([ar_info[self.arm_id]["x"],ar_info[self.arm_id]["y"],ar_info[self.arm_id]["z"]])
             else:
-                marker_1 = np.array([0.002157,0.008755,0.18084])
+                marker_1 = np.array([0.0353238,0.00329190,0.15313373])
         else:
-            marker_1 = np.array([0.002157,0.008755,0.18084])
+            marker_1 = np.array([0.003606,-0.015277,0.138732])
         vec, distance = self.__targetting(marker_1,goal_point)
+        #print(f"vec:{vec[2]}")
         vec[2] = self.calc_t_distance(id,ar_info, vec, distance)
-        goal_area = {"x":[-0.005,0.005],"z":[-0.005,0.005]}
+        goal_area = {"x":[-0.004,0.004],"z":[-0.004,0.004]}
         print(f"distance:{distance},vec:{vec}")
 
         return vec,goal_area
@@ -86,7 +90,7 @@ class ARPowerPlanner():
                 # When z is satisfying the thresholds, cansat changes just orientation
                 motor_ouput = self.STANDARD_POWER - self.POWER_RANGE
                 if vec[0] > goal_area["x"][0] and vec[0] < goal_area["x"][1]:
-                    print("finish")
+                    print("Approach Finished")
                     power_R = 0
                     power_L = 0
                     self.aprc_state = True
@@ -115,11 +119,32 @@ class ARPowerPlanner():
         return {"R":power_R,"L":power_L,"aprc_state":self.aprc_state,"move":move}
 
     def calc_t_distance(self,id,ar_info, vec, distance):
-        y_m = self.rot_vec(ar_info[id]['roll'],ar_info[id]['pitch'],ar_info[id]['yaw'],[0,1,0])
-        vec_normalize = vec.reshape(3,1)/np.linalg.norm(vec)
-        cos_argment = np.dot(y_m.T,vec_normalize)
+        if id == "2" or id == "3" or id == "68": # 68は裏面のマーカー、青モジュールに追加するマーカーも必要
+            y_m = self.rot_vec(ar_info[id]['rvec'],[0,0,1])
+        else:
+            y_m = self.rot_vec(ar_info[id]['rvec'],[0,1,0])
+        vec_normalize = vec.reshape(3,1)/np.linalg.norm(vec[1:3])
+        #print(vec_normalize)
+        cos_argment = np.dot(y_m[1:3].T,vec_normalize[1:3])
+        #print(cos_argment)
         ultraman = distance*np.sqrt(1-cos_argment**2)
-        return ultraman
+        #ultraman = distance*sin(arccos(y_m[1:3].T/vec_normalize[1:3]))
+        return ultraman[0][0]
+        
+    def rotation_matrix(self, axis, theta):
+        """
+        Return the rotation matrix associated with counterclockwise rotation about
+        the given axis by theta radians.
+        """
+        axis = np.asarray(axis)
+        axis = axis / np.sqrt(np.dot(axis, axis))
+        a = np.cos(theta / 2.0)
+        b, c, d = -axis * np.sin(theta / 2.0)
+        aa, bb, cc, dd = a * a, b * b, c * c, d * d
+        bc, ad, ac, ab, bd, cd = b * c, a * d, a * c, a * b, b * d, c * d
+        return np.array([[aa + bb - cc - dd, 2 * (bc + ad), 2 * (bd - ac)],
+                         [2 * (bc - ad), aa + cc - bb - dd, 2 * (cd + ab)],
+                         [2 * (bd + ac), 2 * (cd - ab), aa + dd - bb - cc]])
 
 
     def __targetting(self,marker_1:np.ndarray=np.zeros(3), marker_2:np.ndarray=np.zeros(3)):
@@ -128,9 +153,9 @@ class ARPowerPlanner():
         '''
         target_vec = marker_2 - marker_1
         target_vec = target_vec[0]
-        #print(target_vec)
-        # distance = (self.target_vec[2]/abs(target_vec[2]))*((target_vec[1]**2 + target_vec[2]**2)**0.5)
-        distance = np.sign(target_vec[2])*np.linalg.norm(target_vec[1:2])
+        # print(np.linalg.norm(target_vec))
+        #distance = (self.target_vec[2]/abs(target_vec[2]))*((target_vec[1]**2 + target_vec[2]**2)**0.5)
+        distance = np.sign(target_vec[2])*np.linalg.norm(target_vec[1:3])
         return target_vec, distance
