+++ EtoE/wolvez2023_pkg/Wolvez2023/__init__.py
@@ -5,8 +5,7 @@ print("\n__Importing Wolvez2023 package ...__\n")
 
 from .libcam_module import Picam
 from .ar_module import Target
-from .AR_powerplanner import ARPowerPlanner
-from .cl_powerplanner import ColorPowerPlanner
+from .motor_power_planner import ARPowerPlanner, ColorPowerPlanner
 
 from .arm import Arm
 from .gps import GPS
@@ -14,9 +13,6 @@ from .bno055 import BNO055
 from .lora import lora
 from .motor import Motor
 from .led import led
-
-from .lora_setting import LoraSettingClass
-from .micropyGPS import MicropyGPS
 
 
 __all__ = [
