+++ EtoE/wolvez2023_pkg/Wolvez2023/ar_module.py
@@ -11,7 +11,6 @@ from typing import Union
 import sys
 import time
 import datetime
-import os
 
 arm_id = "1"
 
@@ -29,9 +28,8 @@ class Ar_cansat():
     id_set = [1,2,3,4,5,6,7,8,9,10]
     def __init__(self):
         #レンズの性質などの内部パラメータ(今回はすでに行っている)
-        cwd = os.getcwd()
-        self.camera_matrix = np.load(cwd + "/wolvez2023_pkg/Wolvez2023/mtx.npy")
-        self.distortion_coeff = np.load(cwd + "/wolvez2023_pkg/Wolvez2023/dist.npy")
+        self.camera_matrix = np.load("wolvez2023_pkg/Wolvez2023/mtx.npy")
+        self.distortion_coeff = np.load("wolvez2023_pkg/Wolvez2023/dist.npy")
         # マーカーの辞書選択
         self.dictionary = aruco.getPredefinedDictionary(aruco.DICT_ARUCO_ORIGINAL)  #ARマーカーの生成に使用
         #aruco.customDictionary(nMakers(ID数),Markersize,baseDictionary(基本となるディクショナリ))で独自のディクショナリ作成も可能
@@ -113,7 +111,7 @@ class Ar_cansat():
                     rvec_matrix = rvec_matrix[0] # rodoriguesから抜き出し
                     # 並進ベクトルの転置
                     transpose_tvec = tvec[np.newaxis, :].T
-                    # 合成
+                    # 合成（通称外部パラメータと呼ばれる，回転行列と並進ベクトルを列方向に結合）
                     proj_matrix = np.hstack((rvec_matrix, transpose_tvec))
                     # オイラー角への変換
                     euler_angle = cv2.decomposeProjectionMatrix(proj_matrix)[6] # [deg]
@@ -165,7 +163,7 @@ class Ar_cansat():
                     #cv2.imshow('drawDetectedMarkers', img)
                     #cv2.waitKey(0)
                     #cv2.destroyAllWindows()
-                    self.ar_info[str(i)] = {'x':tvec[0],'y':tvec[1],'z':tvec[2],'roll':euler[0],'pitch':euler[1],'yaw':euler[2],'norm':self.norm_tvec}
+                    self.ar_info[str(i)] = {'x':tvec[0],'y':tvec[1],'z':tvec[2],'roll':euler[0],'pitch':euler[1],'yaw':euler[2],'norm':self.norm_tvec,'rvec':rvec}
                     # self.ar_info.append(info)
                     
             
@@ -186,41 +184,43 @@ class Ar_cansat():
         if connecting_state == 0:
             if "2" in ar_info.keys():
                 self.aprc_AR = True
-                side = 'marker_R'
                 norm = ar_info['2']['norm']
                 target_id = '2'
             else: 
-                side = 'None'
                 norm = 0
                 self.aprc_AR = False
                 target_id = 100
 
         else:
-            if "3" in ar_info.keys() or "4" in ar_info.keys():
+            if "2" in ar_info.keys():
                 self.aprc_AR = True
-                side = 'marker_R'
-                if "3" in ar_info.keys():
-                    norm = ar_info['3']['norm']
-                    target_id = '3'
-                else:
-                    norm = ar_info['4']['norm']
-                    target_id = '4'
-            elif "5" in ar_info.keys() or "6" in ar_info.keys():
+                norm = ar_info['2']['norm']
+                target_id = '2'
+            elif "5" in ar_info.keys() or "6" in ar_info.keys() or "7" in ar_info.keys(): # マーカー追加予定(id:68)
                 self.aprc_AR = True
-                side = 'marker_L'
                 if "5" in ar_info.keys():
                     norm = ar_info['5']['norm']
                     target_id = '5'
-                else:
+                elif "6" in ar_info.keys():
                     norm = ar_info['6']['norm']
                     target_id = '6'
+                else:
+                    norm = ar_info['7']['norm']
+                    target_id = '7'
+            elif "3" in ar_info.keys() or "4" in ar_info.keys():
+                self.aprc_AR = True
+                if "4" in ar_info.keys():
+                    norm = ar_info['4']['norm']
+                    target_id = '4'
+                else:
+                    norm = ar_info['3']['norm']
+                    target_id = '3'
             else:
                 norm = 0
-                side = 'None'
                 self.aprc_AR = False
                 target_id = 100
         
-        return {"AR":self.aprc_AR, "side":side, "id":target_id, "norm":norm}
+        return {"AR":self.aprc_AR, "id":target_id, "norm":norm}
 
 
 class Target(Ar_cansat):
