+++ EtoE/wolvez2023_pkg/Wolvez2023/arm.py
@@ -16,7 +16,7 @@ class Arm():
 		
 	def up(self,buff=0):
 		self.pi.write(self.servo_pin,1)
-		self.pi.set_servo_pulsewidth(self.servo_pin,1500)
+		self.pi.set_servo_pulsewidth(self.servo_pin,1650)
 		time.sleep(0.5)
 		self.pi.set_servo_pulsewidth(self.servo_pin,0)
 		self.pi.write(self.servo_pin,0)
@@ -30,7 +30,7 @@ class Arm():
 		
 	def down(self,buff=0):
 		self.pi.write(self.servo_pin,1)
-		self.pi.set_servo_pulsewidth(self.servo_pin,1150)
+		self.pi.set_servo_pulsewidth(self.servo_pin,1000)
 		time.sleep(0.5)
 		self.pi.set_servo_pulsewidth(self.servo_pin,0)
 		self.pi.write(self.servo_pin,0)
@@ -81,9 +81,11 @@ if __name__ == "__main__":
 	#arm = ArmHardPwm(16)
 	arm = Arm(23)
 	arm.setup()
-	arm.move(850)
-	arm.move(1800)
+	arm.up()
+	#arm.move(1800)
+	#arm.middle()
 	arm.down()
+	arm.middle()
 	start = time.time()
 	while True:
 		end = time.time()
