+++ EtoE/wolvez2023_pkg/Wolvez2023/libcam_module.py
@@ -7,7 +7,7 @@ from datetime import datetime
 class Picam():
     def __init__(self):
         # define camera parameter
-        self.size = (1800, 1000)
+        self.size = (1750, 1300)
 
         # setting picam2 up
         self.picam2 = Picamera2()
@@ -22,7 +22,26 @@ class Picam():
 
         # Libcamera's setting to use AF mode (AFSpeed Fast)
         # self.picam2.set_controls({"AfMode": controls.AfModeEnum.Continuous,"AfSpeed":controls.AfSpeedEnum.Fast})
-        self.picam2.set_controls({"AfMode":0,"LensPosition":4.5})
+        self.picam2.set_controls({"AfMode" : 0, "LensPosition" : 9})
+        
+    def change_size(self,height, width, lensposition=9.5):
+        self.size - (height,width)
+        
+        self.stop()
+        
+        # setting picam2 up
+        self.picam2 = Picamera2()
+        config = self.picam2.create_preview_configuration(
+            main={"format": 'XRGB8888', "size": self.size})
+        self.picam2.align_configuration(config)
+        self.picam2.configure(config)
+        self.picam2.start()
+
+        # Libcamera's setting to use AF mode
+        self.picam2.set_controls({"AfMode": controls.AfModeEnum.Continuous})
+
+        # Libcamera's setting to use AF mode (AFSpeed Fast)
+        self.picam2.set_controls({"AfMode" : 0, "LensPosition" : lensposition})
         
     def capture(self, args, filename="test"):
         self.filename = filename
